pyproject.toml
[project]
name = "bergen-smart-rain-hub"
version = "0.1.0"
description = "Analyse emergency rainwater supply potential in Bergen"
requires-python = ">=3.11"
license = "Apache-2.0"

dependencies = [
    "streamlit>=1.55,<2",
    "pandas>=2.3,<3",
    "numpy>=2.4,<3",
    "altair>=6,<7",
    "requests>=2.32,<3",
    "python-dotenv>=1.2,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
